Library Manager compatible folder structure
The Arduino IDE Library Manager does not support libraries that are in
subfolders. This also makes the library compatible with manual
installation via Sketch > Include Library > Add .ZIP Library in the
Arduino IDE'.
